+++ packages/client/package.json
@@ -15,6 +15,7 @@
         "build": "tsup src/index.ts --format cjs,esm --dts"
     },
     "devDependencies": {
+        "@types/ws": "^8.5.10",
         "tsup": "^8.1.0"
     }
 }
