+++ packages/server/src/core/index.ts
@@ -1,10 +1,10 @@
+export * from "./aetheris";
 export * from "./aetheris-error";
+export * from "./cookie-manager";
 export * from "./create-aetheris";
 export * from "./create-router-map";
 export * from "./get-match";
 export * from "./handler";
 export * from "./middleware";
-export * from "./procedure";
-export * from "./cookie-manager";
 export * from "./procedure-response";
 export * from "./router";
